bad bad

+++ LearnOpenGL/main.cpp
@@ -112,7 +112,7 @@ static void draw()
 	// update uniform
 	//int vertexColorLocation = glGetUniformLocation(shader.getShaderProgram(), "xyoffset");
 	//glUniform2f(vertexColorLocation, sin(glfwGetTime()), cos(glfwGetTime()));
-	shader.write("xyoffset", { sin(glfwGetTime()) ,cos(glfwGetTime()) });
+	shader.write("xyoffset", { static_cast<float>(sin(glfwGetTime())) ,static_cast<float>(cos(glfwGetTime())) });
 
 	glActiveTexture(GL_TEXTURE0);
 	glBindTexture(GL_TEXTURE_2D, texture[0]);
